Update Chinese Simplified Translate
"function" is officially translated as "方法" when it means a method in an object (or a class).
"closing bracket" here means the bracket that matches the current one so it should be translated as "配对括号" instead of "关闭括号" which means "close the bracket".
"debug script" should be "调试脚本". The current one has wrong word order. 
Add extra information to "Line_Case_Insensitive" and "Lines" to help users understand.
"move line down" should be "下移行". The current one has wrong word order. 
"move line up" should be "上移行". The current one has wrong word order.
"save to disk" is not modified. Actually the function of this button is to save the script to local file system instead of disk. However the English translation said "disk".
"scope" is officially translated as "块" when it means the texts between "{" and "}".
"select between brackets" is changed to "选择前后括号中间的部分" because the current one "选择括号间" is strange.
"selection" should be "选择" instead of "选择内容" which means "select content".
"toggle" here means to add or remove something, not to switch.
"incremental" means "增量" instead of "增亮".