Update messages.json – thorough review and update
I have made a thorough review and update of the Danish translation.

I corrected a lot of spelling and typing mistakes, mostly regarding use of lower and upper case and compound nouns. We don't use nearly as many capital letters in Danish as in English, and compound nouns are always in one word in Danish.

I also added and translated quite a few missing entries.